feature: implemented the link and unlink actions
